utils/cryptography.py
 
# ============LICENSE_START========================================== 
# org.onap.vvp/test-engine
# ===================================================================
# Copyright © 2017 AT&T Intellectual Property. All rights reserved.
# ===================================================================
#
# Unless otherwise specified, all software contained herein is licensed
# under the Apache License, Version 2.0 (the “License”);
# you may not use this software except in compliance with the License.
# You may obtain a copy of the License at
#
#             http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.
#
#
#
# Unless otherwise specified, all documentation contained herein is licensed
# under the Creative Commons License, Attribution 4.0 Intl. (the “License”);
# you may not use this documentation except in compliance with the License.
# You may obtain a copy of the License at
#
#             https://creativecommons.org/licenses/by/4.0/
#
# Unless required by applicable law or agreed to in writing, documentation
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.
#
# ============LICENSE_END============================================
#
# ECOMP is a trademark and service mark of AT&T Intellectual Property.
from itsdangerous import URLSafeTimedSerializer
from rest_framework_jwt.settings import api_settings

from services.logging_service import LoggingServiceFactory


logger = LoggingServiceFactory.get_logger()

class CryptographyText(object):

    @staticmethod
    def encrypt(text):
        encryptor = URLSafeTimedSerializer(api_settings.JWT_SECRET_KEY)
        return encryptor.dumps(text, salt=api_settings.JWT_SECRET_KEY)

    @staticmethod
    def decrypt(encoded_text):
        decryptor = URLSafeTimedSerializer(api_settings.JWT_SECRET_KEY)
        text = decryptor.loads(
            encoded_text,
            salt=api_settings.JWT_SECRET_KEY,
        )

        return text
